+++ patching/plugin-commands-patching/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pnpm/plugin-commands-patching
 
+## 3.1.15
+
+### Patch Changes
+
+- @pnpm/plugin-commands-installation@13.0.20
+- @pnpm/store-connection-manager@6.0.20
+- @pnpm/config@18.4.4
+- @pnpm/patching.apply-patch@2.0.4
+
+## 3.1.14
+
+### Patch Changes
+
+- @pnpm/store-connection-manager@6.0.19
+- @pnpm/plugin-commands-installation@13.0.19
+
 ## 3.1.13
 
 ### Patch Changes