+++ pkg-manager/plugin-commands-installation/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @pnpm/plugin-commands-installation
 
+## 13.0.20
+
+### Patch Changes
+
+- @pnpm/plugin-commands-rebuild@9.1.3
+- @pnpm/core@10.2.10
+- @pnpm/package-store@16.0.9
+- @pnpm/store-connection-manager@6.0.20
+- @pnpm/config@18.4.4
+
+## 13.0.19
+
+### Patch Changes
+
+- @pnpm/core@10.2.9
+- @pnpm/outdated@13.0.8
+- @pnpm/package-store@16.0.8
+- @pnpm/store-connection-manager@6.0.19
+- @pnpm/plugin-commands-rebuild@9.1.2
+
 ## 13.0.18
 
 ### Patch Changes