+++ store/plugin-commands-store/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pnpm/plugin-commands-store
 
+## 8.0.10
+
+### Patch Changes
+
+- Updated dependencies [4bbf482d1]
+  - @pnpm/store.cafs@1.0.0
+  - @pnpm/store-connection-manager@6.0.20
+  - @pnpm/config@18.4.4
+
+## 8.0.9
+
+### Patch Changes
+
+- @pnpm/store-connection-manager@6.0.19
+
 ## 8.0.8
 
 ### Patch Changes